LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-expose-star-worker-expected.txt


FAIL Basic Access-Control-Expose-Headers: * support assert_equals: expected (string) "X" but got (object) null
PASS Cannot use * for credentialed fetches 
